# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grainily API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/modes/{key}:
    put:
      operationId: adminUpdateMode
      tags: [admin]
      summary: Admin - toggle a platform mode
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformModeInput"
      responses:
        "200":
          description: Updated mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformMode"
...
  /admin/members:
    get:
      operationId: adminListMembers
      tags: [admin]
      summary: Admin - list regular users (customers/merchants)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, merchant, moderator]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberUser"
    post:
      operationId: adminCreateMember
      tags: [admin]
      summary: Admin - create a regular user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUserInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberUser"
  /admin/members/{id}:
    put:
      operationId: adminUpdateMember
      tags: [admin]
      summary: Admin - update a regular user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUserInput"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberUser"
    delete:
      operationId: adminDeleteMember
      tags: [admin]
      summary: Admin - delete a regular user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/members/{id}/suspend:
    patch:
      operationId: adminSuspendMember
      tags: [admin]
      summary: Root - suspend a user account, blocking login and active sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberUser"
  /admin/members/{id}/unsuspend:
    patch:
      operationId: adminUnsuspendMember
      tags: [admin]
      summary: Root - reactivate a suspended user account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberUser"
  /admin/members/{id}/impersonate:
    post:
      operationId: adminImpersonateMember
      tags: [admin]
      summary: Root - issue a login token for a member to browse the app as them
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Impersonation token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpersonationResult"
  /admin/audit-log:
    get:
      operationId: adminListAuditLog
      tags: [admin]
      summary: Root - list/filter the audit log of admin & root actions
      security:
        - bearerAuth: []
      parameters:
        - name: actorId
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: targetType
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogList"
  /admin/audit-log/actions:
    get:
      operationId: adminListAuditLogActions
      tags: [admin]
      summary: Root - list distinct action types recorded in the audit log
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Distinct action types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
components:
  schemas:
...
    PlatformMode:
      type: object
      required: [key, label, enabled]
      properties:
        key:
          type: string
        label:
          type: string
        description:
          type: ["string", "null"]
        enabled:
          type: boolean
    PlatformModeInput:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
...
    MemberUser:
      type: object
      required: [id, name, email, role, createdAt, emailVerified, isSuspended]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, merchant, moderator]
        phone:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]
        isSuspended:
          type: boolean
    ImpersonationResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          type: object
          required: [id, name, email, role]
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
            role:
              type: string
    MemberUserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [customer, merchant, moderator]
        phone:
          type: string
        emailVerified:
          type: boolean
...
    AuditLogEntry:
      type: object
      required: [id, actorId, actorRole, actorLabel, action, targetType, summary, createdAt]
      properties:
        id:
          type: integer
        actorId:
          type: string
        actorRole:
          type: string
          enum: [root, admin]
        actorLabel:
          type: string
        action:
          type: string
        targetType:
          type: string
        targetId:
          type: ["string", "null"]
        summary:
          type: string
        metadata:
          nullable: true
        createdAt:
          type: string
    AuditLogList:
      type: object
      required: [rows, total, page, limit]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT