# lib/api-zod/src/generated/types/auditLogList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grainily API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogEntry } from './auditLogEntry';

export interface AuditLogList {
  rows: AuditLogEntry[];
  total: number;
  page: number;
  limit: number;
}